backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
factory_boy==3.3.1
Faker==30.3.0
idna==3.10
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
